decompose_logprices

Seasonal and Trend decomposition using Loess (STL) decomposition of historical wholesale prices.
A log-type transformation $$y=\operatorname{arcsinh}(\frac{x}{scale})$$ This is referred to as the "logprice" and is close to a log transformation when x is larger than \(scale\). However, it crosses over to a linear function for small xand therefore handle negative wholesale prices smoothly.
The seasonal decomposition is in to periodic daily, weekly and aperiodic yearly terms. decompose_logprices() also fills any gaps in the input data to form an hourly time series.

The output of decompose_logprices() is the key input to generate_logprice_hmm() and to simulate_prices()

decompose_logprices(price_data)

Arguments

price_data

price data in format datetime,price e.g. sem_prices_2019_2025

Value

dataframe with STL decomposition of hourly price data

Examples

decompose_logprices(sem_prices_2019_2025)
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> # A tibble: 61,368 × 8
#>    datetime            logprice trend season_day season_week season_year
#>    <dttm>                 <dbl> <dbl>      <dbl>       <dbl>       <dbl>
#>  1 2019-01-01 00:00:00     9.55  9.43    -0.157    -0.0213        0.0645
#>  2 2019-01-01 01:00:00     9.48  9.43    -0.183    -0.0117        0.0645
#>  3 2019-01-01 02:00:00     9.41  9.43    -0.212    -0.00306       0.0645
#>  4 2019-01-01 03:00:00     9.48  9.43    -0.215     0.000605      0.0645
#>  5 2019-01-01 04:00:00     9.48  9.43    -0.187     0.0173        0.0645
#>  6 2019-01-01 05:00:00     9.52  9.43    -0.105     0.0430        0.0645
#>  7 2019-01-01 06:00:00     9.44  9.43     0.0111    0.0664        0.0645
#>  8 2019-01-01 07:00:00     9.24  9.43     0.115     0.0806        0.0645
#>  9 2019-01-01 08:00:00     9.24  9.43     0.129     0.0601        0.0645
#> 10 2019-01-01 09:00:00     9.39  9.43     0.104     0.0483        0.0645
#> # ℹ 61,358 more rows
#> # ℹ 2 more variables: remainder <dbl>, annual_leakage <dbl>