evaluate_tariffs

Agents' prospect-theory evaluation of the initial fixed to flexible switching decision. evaluate_tariffs() uses get_prospect_costs() (the five annual costs) and get_ce_value() (the certainty-equivalent of each plan), then decides to adopt whichever of tou dynamic has the most favourable certainty-equivalent gain. If there is no gain, the agents stay on flat. This is the self-contained unit meant to be called once per household from update_agents() in place of the classic savings/social/theta decision.

evaluate_tariffs(
  scen,
  kWh,
  phi,
  gamma,
  eta,
  tau,
  natural_profile = "LP1",
  smart_rollout,
  dyn_aware,
  flex_certainty_tou = 0.8,
  flex_certainty_det = 0.7,
  lambda = 2.25,
  prices_scen,
  params
)

Arguments

scen

scenario dataframe

kWh

annual load

phi

\(\phi\)

gamma

\(\gamma\)

eta

\(\eta\)

tau

\(\tau\)

natural_profile

the characteristic profile of the household (currently LP1 or LP3)

smart_rollout

smart meter rollout time

dyn_aware

dynamic tariff awareness time

flex_certainty_tou

confidence in tou flexibility

flex_certainty_det

confidence in det flexibility

lambda

loss aversion

prices_scen

price scenario

params

parameters at params$yeartime

Value

a list: costs (the five annual costs), ce (the full get_ce_value()

output), and decision ("flat", "tou", or "dynamic")

Examples

prices_scen <- set_prices(sD)
params <- scenario_params(sD,2032)
evaluate_tariffs(sD,kWh = 8760, phi = 0.4, gamma = 5, eta = 0.3, tau = 40,natural_profile = "LP1", smart_rollout = 2025,dyn_aware=2028,0.6,0.5,10.25,prices_scen = prices_scen,params)
#> $costs
#> # A tibble: 1 × 5
#>    flat tou_noflex tou_flex det_noflex det_flex
#>   <dbl>      <dbl>    <dbl>      <dbl>    <dbl>
#> 1 3363.      3425.    2938.      3481.    2955.
#> 
#> $ce
#> $ce$flat
#> [1] 3362.634
#> 
#> $ce$certainty_tou
#> [1] 0.6
#> 
#> $ce$CE_tou
#> [1] -6.852908
#> 
#> $ce$CE_bill_tou
#> [1] 3369.487
#> 
#> $ce$certainty_det
#> [1] 0.5
#> 
#> $ce$CE_det
#> [1] -34.02333
#> 
#> $ce$CE_bill_det
#> [1] 3396.658
#> 
#> 
#> $decision
#> [1] "flat"
#>