generate_logprice_hmm

creates an n state hidden Markov model based on residuals extracted from decompose_logprices. The default is three states, supposed to represent high RES/low demand, normal RES/demand and low RES/high demand regimes.

generate_logprice_hmm uses a gaussian model at present. In principle should be replaced with fat-tailed distribution e.g. student-t in future because outliers lead to an overestimation of standard deviation. The workaround used at present is to winsorise the logprice residuals.

It is not necessary to run this function at the beginning of each simulation run. The 2019-2025 calibration is in the model hmm_fit.

generate_logprice_hmm(dcmp, n_states = 3, winsor = 0.01)

Arguments

dcmp

decomposed logprices from decompose_logprices()

n_states

number of Markov states, default 3

winsor

winsorisation (clipping) level e.g if winsor=0.01 quantile bounds are 1-99

Value

a depmixS4 model object

Examples

dcmp <- decompose_logprices(sem_prices_2019_2025)
generate_logprice_hmm(dcmp,n_state=3)
#> iter = 1 
#> LL = -124011.0410112 
#> diff = Inf 
#> 
#> iter = 2 
#> LL = -60133.8893174 
#> diff = 63877.15 
#> 
#> iter = 3 
#> LL = -14080.7405909 
#> diff = 46053.15 
#> 
#> iter = 4 
#> LL = -8255.3400740 
#> diff = 5825.401 
#> 
#> iter = 5 
#> LL = -6702.4585955 
#> diff = 1552.881 
#> 
#> iter = 6 
#> LL = -6140.1140061 
#> diff = 562.3446 
#> 
#> iter = 7 
#> LL = -5815.3869679 
#> diff = 324.727 
#> 
#> iter = 8 
#> LL = -5594.9899785 
#> diff = 220.397 
#> 
#> iter = 9 
#> LL = -5442.1128062 
#> diff = 152.8772 
#> 
#> iter = 10 
#> LL = -5339.1508792 
#> diff = 102.9619 
#> 
#> iter = 11 
#> LL = -5272.1751208 
#> diff = 66.97576 
#> 
#> iter = 12 
#> LL = -5228.3887315 
#> diff = 43.78639 
#> 
#> iter = 13 
#> LL = -5198.8404507 
#> diff = 29.54828 
#> 
#> iter = 14 
#> LL = -5178.2783315 
#> diff = 20.56212 
#> 
#> iter = 15 
#> LL = -5163.6871751 
#> diff = 14.59116 
#> 
#> iter = 16 
#> LL = -5153.2384550 
#> diff = 10.44872 
#> 
#> iter = 17 
#> LL = -5145.7400247 
#> diff = 7.49843 
#> 
#> iter = 18 
#> LL = -5140.3652579 
#> diff = 5.374767 
#> 
#> iter = 19 
#> LL = -5136.5204896 
#> diff = 3.844768 
#> 
#> iter = 20 
#> LL = -5133.7745682 
#> diff = 2.745921 
#> 
#> iter = 21 
#> LL = -5131.8149235 
#> diff = 1.959645 
#> 
#> iter = 22 
#> LL = -5130.4163259 
#> diff = 1.398598 
#> 
#> iter = 23 
#> LL = -5129.4174561 
#> diff = 0.9988697 
#> 
#> iter = 24 
#> LL = -5128.7032771 
#> diff = 0.714179 
#> 
#> iter = 25 
#> LL = -5128.1919619 
#> diff = 0.5113152 
#> 
#> iter = 26 
#> LL = -5127.8253604 
#> diff = 0.3666015 
#> 
#> iter = 27 
#> LL = -5127.5621381 
#> diff = 0.2632224 
#> 
#> iter = 28 
#> LL = -5127.3728832 
#> diff = 0.1892549 
#> 
#> iter = 29 
#> LL = -5127.2366369 
#> diff = 0.1362463 
#> 
#> iter = 30 
#> LL = -5127.1384380 
#> diff = 0.0981989 
#> 
#> iter = 31 
#> LL = -5127.0675878 
#> diff = 0.07085017 
#> 
#> iter = 32 
#> LL = -5127.0164223 
#> diff = 0.05116549 
#> 
#> iter = 33 
#> LL = -5126.9794422 
#> diff = 0.03698008 
#> 
#> iter = 34 
#> LL = -5126.9526956 
#> diff = 0.02674663 
#> 
#> iter = 35 
#> LL = -5126.9333385 
#> diff = 0.01935715 
#> 
#> iter = 36 
#> LL = -5126.9193216 
#> diff = 0.01401683 
#> 
#> iter = 37 
#> LL = -5126.9091670 
#> diff = 0.0101546 
#> 
#> iter = 38 
#> LL = -5126.9018074 
#> diff = 0.007359595 
#> 
#> iter = 39 
#> LL = -5126.8964716 
#> diff = 0.005335794 
#> 
#> iter = 40 
#> LL = -5126.8926019 
#> diff = 0.003869705 
#> 
#> iter = 41 
#> LL = -5126.8897947 
#> diff = 0.002807195 
#> 
#> iter = 42 
#> LL = -5126.8877578 
#> diff = 0.002036893 
#> 
#> iter = 43 
#> LL = -5126.8862796 
#> diff = 0.001478262 
#> 
#> iter = 44 
#> LL = -5126.8852066 
#> diff = 0.001073029 
#> 
#> iter = 45 
#> LL = -5126.8844276 
#> diff = 0.0007790019 
#> 
#> iter = 46 
#> LL = -5126.8838619 
#> diff = 0.00056562 
#> 
#> iter = 47 
#> LL = -5126.8834512 
#> diff = 0.0004107367 
#> 
#> iter = 48 
#> LL = -5126.8831529 
#> diff = 0.0002982969 
#> 
#> iter = 49 
#> LL = -5126.8829362 
#> diff = 0.0002166586 
#> 
#> iter = 50 
#> LL = -5126.8827789 
#> diff = 0.0001573766 
#> 
#> iter = 51 
#> LL = -5126.8826645 
#> diff = 0.0001143247 
#> 
#> iter = 52 
#> LL = -5126.8825815 
#> diff = 8.305627e-05 
#> 
#> iter = 53 
#> LL = -5126.8825211 
#> diff = 6.034356e-05 
#> 
#> iter = 54 
#> LL = -5126.8824773 
#> diff = 4.384521e-05 
#> 
#> iter = 55 
#> LL = -5126.8824454 
#> diff = 3.185903e-05 
#> 
#> iter = 56 
#> LL = -5126.8824223 
#> diff = 2.315149e-05 
#> 
#> iter = 57 
#> LL = -5126.8824055 
#> diff = 1.682455e-05 
#> 
#> iter = 58 
#> LL = -5126.8823932 
#> diff = 1.22273e-05 
#> 
#> iter = 59 
#> LL = -5126.8823843 
#> diff = 8.886837e-06 
#> 
#> $pi_mat
#>              [,1]       [,2]         [,3]
#> [1,] 8.732982e-01 0.12670179 8.101153e-10
#> [2,] 1.433031e-02 0.91448017 7.118952e-02
#> [3,] 7.934285e-23 0.04127338 9.587266e-01
#> 
#> $delta
#> [1] 6.877101e-210  2.525334e-63  1.000000e+00
#> 
#> $mu
#> [1] -7.12488204  0.13244397  0.07911517
#> 
#> $sigma
#> [1] 5.5182553 0.4410306 0.1285604
#>