get_annual_cost_simple

evaluates the annual electricity cost of tariff scheme at yeartime for an agent based on standard profiles i.e. no flexibility modelling. This function is used by initialise_agents() and is also useful
At present the tariff scheme are flat, day/night/peak and dynamic and the characteristic load profile is LP1 or LP3. Future price assumptions are typically the output of get_tariff_prices(scenario)

get_annual_cost_simple(
  yeartime,
  kWh,
  tariff_plan,
  profile = "LP1",
  prices_scen
)

Arguments

yeartime

decimal time

kWh

annual consumption kWh (assumed inflexible)

tariff_plan

flat, tou or dynamic

profile

household usage profile e.g. LP1

prices_scen

tariff price scenario

Value

a real number, euros

Examples


prices_scen <- set_prices(sD)
get_annual_cost_simple(2019,4000,"tou_old","LP2",prices_scen=prices_scen)
#> [1] 863.9081
get_annual_cost_simple(2030,8760,"tou_old","LP2",prices_scen=prices_scen)
#> [1] 2707.416
get_annual_cost_simple(2030,8760,"tou","LP1",prices_scen=prices_scen)
#> [1] 2894.575
get_annual_cost_simple(2030,8760,"flat","LP1",prices_scen=prices_scen)
#> [1] 2833.363
get_annual_cost_simple(2030,8760,"dynamic","LP1",prices_scen=prices_scen)
#> [1] 2862.518