get_flex

get_flex solves for the flexible hourly household load pattern \(L_t\) that optimises the cost to a household over the hourly period 1:T. Here, the "cost" represents the direct billing cost as well as penalties associated with load-shifting behaviour. get_flex() assumes that total consumption over the period 1:T is unchanged irrespective of prices \(p(t)\)

The "natural" household load profile without load-shifting is \(L_t^0\), corresponding to a flat electricity price profile of the household. Load-shifting arises in response to time-varying prices \(p_t\). The billing cost is is linear \(\sum_{t=1}^T p_t L_t\) where \(L_t\) is the load-shifted hourly consumption. Two additional cost penalties arise: $$\delta c = \sum_{t=1}^T p_t x_t + \eta \sum_t (x_{t+1}-x_t)^2 + \gamma \sum_t \left( \sum_k g(k-t) x_k \right)^2$$
The quantity to be minimised \(\cal{c}=\frac{C}{E})\) where \(E=\sum_t L_t = \sum_t L_t^0\). \(x_t\) is the dimensionless household load profile, and \(\eta,\gamma\) are cost parameters with dimensions of euro per kWh.

get_flex(
  demand,
  phi = 0.4,
  gamma = 5,
  eta = 0.3,
  tau = 24,
  kernel = "exp",
  precision = 1e-04
)

Arguments

demand

a 3 column dataframe of datetime, hourly prices and natural_load

phi

baseload (inflexible) fraction of load

gamma

unscaled quandatric cost penalty weight

eta

unscaled weight of the kinetic term (regularisation)

tau

flexibility time horizon in hours

kernel

choice of "matern", "exp", "cauchy","gauss" (both have single lifetime tau)

precision

desired solver precision

Value

a 7-column dataframe of datetime, price, unperturbed load, flexible share of unperturbed load, load-shift x, optimised load, and inflexible load

Details

The second term is a "ramping" cost penalty that limits unrealistic rapid household load variations in response to minor price variations and controls the overall smoothness of the \(x_t\). The third term is the square of a convolution of \(x_t\) with a kernel \(g_t\). \(g_t\) decays on some timescale \(\tau\) that represents the time horizon over which households are willing to defer or advance consumption in response to anticipated price changes. It is natural to normalise the kernel so that \(\int g_t^2 dt = 1\). The standard symetric choice in \(depmicrosimr\) is \(g(t)= \sqrt{\frac{1}{\tau}} \exp(-|t|/\tau)\).

The \(\eta\) term prevents excessive load-shifting in response to small differences in prices (regularisation). The \(\gamma_\tau\) term is permits is a stiffness term that limits load-shifting over a timescale \(\tau\) but inhibits it on longer timescales.

The cost parameters are defined for a standard annual load of 8760kWh.

get_flex() currently uses a simple exponential kernel (but other choices are possible)

Three additional constraints are imposed (1) \(\sum(x_t)=0\) is i.e. total household energy consumption is inelastic (2) the maximum import capacity MIC cannot be exceeded \(L^0_t + \sum(x_t) <= MIC\) (iii) consumption load cannot be negative \(L^0_t + \sum(x_t) >= 0\).

The inflexible fraction \(\phi\) is modulated by a time-of-day factor \(f_t\). This is inferred from the lowest quantile of smart meter dataset. Generally \(f_t > 1\) at night and less than one during the day. The inflexible fraction is defined as $$ \phi_t = \phi-(1-\phi)\log(1-\phi)*(f_t-1)$$ This has the desired properties the mean of \(\phi_t\) equals \(\phi\) and \(\phi_t \rightarrow 1\) as \(\phi \rightarrow 1\), irrespective of \(f_t\).

Depending in user intent, two parameter scaling models are implemented - "

Examples


demand <- set_prices(sD)
demand <- demand %>% dplyr::filter(lubridate::year(datetime)==2026)
demand <- demand %>% dplyr::inner_join(load_profiles_generalised %>% dplyr::select(datetime,lp1))
#> Joining with `by = join_by(datetime, lp1)`
demand <- demand %>% dplyr::mutate(load=4200*lp1) %>% dplyr::select(-lp1)
demand <- demand %>% dplyr::filter(tariff_plan=="tou") %>% dplyr::select(datetime,price,load)
test <- get_flex(demand,phi=0.4,gamma=5,eta=0.3,tau=72)
100*sum(abs(test$load_opt-test$load))/8760
#> [1] 43.98642