get_flex_scores

Evaluates a measure of how much load has been shifted in response a specific price incentive (tariff plan). The flexibility is defined as $$ \frac{1}{2} \frac{\sum_t |L_{optimised}-L_{natural}|}{\sum_t L_{natural}}$$The evaluation period is one year. This is the same as \(\frac{1}{2} \sum_t \mid x_t \mid\) (see get_flex)

The main use of this function is to map out the relationship flexibility parameters and stated flexibility scores.
Based on the LP1 profile (2026 urban flat rate)

get_flex_scores(
  scen,
  year,
  kWh,
  tariff_plan,
  phi,
  gamma,
  eta,
  tau,
  kernel,
  prices_scen
)

Arguments

scen

scenario

year

year of evaluation

kWh

annual consumption

tariff_plan

tariff plan

phi

inflexibility parameter

gamma

cost parameter

eta

kinetic cost parameter

tau

load-shift time horizon

kernel

defaut to "exp"

prices_scen

price scenario

Value

a single row dataframe

Examples

#prices_scen <- set_prices(sD)
#get_flex_scores(sD,2026,8760,"tou",0.7,0,0,48,"exp",prices_scen)
#get_flex_scores(sD,2026,8760,"tou",1,1,0.6,48,"exp",prices_scen)