initialise_agents

initialise_agents() sets the initial state variables at the beginning of each run (Jan 2019 before smart-meter rollout)

Household "stated flexibility" scores are inferred from responses survey questions. These scores are normally distributed around zero. A flexibility score in response to ToU pricing can be defined as the hourly annual sum. $$\frac{1}{2} \sum_t | \hat{l}(t)-\hat{l}^0(t) | $$ where \(\hat{l}(t)\) and \(\hat{l}^0(t)\) are the load-shifted and flat-price ("natural") load profiles. This flexibility score index lies between 0 & 1 and can ve computed using get_flex() using 2026 prices, for example. initialise_agents() maps stated flexibilities to hourly flexibility scores using a Beta distribution in the range \([0,f_{max}]\). This means that hourly flexibility scores are defined for each agent.

The flexibility cost model used by depmicrosimr uses four parameters. These are an inflexible load fraction (\(\phi\)), a ramping cost parameter (\(\eta\)), a load-shifting cost parameter (\(\gamma\)), and a load reversion timescale (\(\tau\)) (see get_flex(). The parameters \(\gamma,\tau\) are set stochastically based on the derived flexibility scores, with \(\phi\) and \(\eta\) are fixed.The latter parameters are adjusted so that the aggregate flexibility matches the difference between 2026 LP1 ("flat") and LP2("ToU") profiles. The hetereogenity in flexibility is described by \(\gamma-\tau\) parameter pair. From the triangular inequality, aggregrate flexibilty is lower than the weighted sum of household flexibilities $$f_{aggregate} < \sum_i^N w_i f_i$$ (weighted by the individual household annual demand). This means that there is a one-to-many mapping between the derived flexibility scores characteristic to each household and the model flexibility parameters.

initialise_agents(
  scen,
  start_year = 2019,
  prices_scen,
  social_network,
  eta = 0.3,
  phi = 0.4
)

Arguments

scen

scenario design dataframe e.g. sD

start_year

default 2019

prices_scen

tariff prices dataframe

social_network

social network

eta

eta flex parameter value

phi

\(\phi\) parameter choice

Value

a dataframe with columns serial ID, annual kWh, initial tariff plan, smart meter install time, and behavioural parameters

Details

Start year (default 2019) is assumed to be before the beginning smart meter rollout. The old day/night dual-metering system ("tou_old") is used by about 12% of households.

The network input is used to determine the social degree of each agent (correlated with but not identical to the stated degree).

Each agent becomes aware of the existence of DEP as a tariff choice at a random time between 2026.5 and dyn_aware_year.

Examples

prices_scen <- set_prices(sD)
social_network <- make_artificial_society(dep_society_1,homophily,nu=4.5)
test <- initialise_agents(sD,2019,prices_scen,social_network,0.3,0.4)
#> [1] "The maximum possible initial flexibility is 46%"
#> [1] "median household flexibilities 22.9 % vs lp1-lp2 flexibility 14.4%"
#> [1] "weighted mean of household flexibilities 23.6"