set_prices

set_prices() is a retail electricity tariff pricing model based on wholesale prices and network charges, and retail supplier pricing assumptions. The purpose of this simple model is to have a consistent set of flat, day/night/peak and dynamic prices in future projections and also to reproduce historic prices with reasonable accuracy. The tariff_plan are supposed to be broadly reflective of the market, rather than to mimic one particular supplier. Without this sub-model, with ad-hoc assumptions for future tariff plan pricing, the ABM projections might not be meaningful because they would depend on specific assumptions.

This is equivalent to the assumption that network supplier set their prices so that the average price paid by customers on flat, ToU and dynamic prices are related across standard profiles.The parameter w represents the fraction of gain from flexibility that is passed on to consumers. If the gain is passed on in full, then \(p_{flat}=p_{tou}\) where the weighted mean prices are calculated in the inflexible LP1 profile. In general however, \(p_{flat} < p_{tou}\) for inflexible customers, and therefore there is a potential cost from adopting variable pricing tariffs. For inflexible customers, set_prices() gives \(p_{tou}\approx p_{dynamic} > p_{flat}\)

set_prices(scen, end_year = 2045, cru_cap = TRUE, shock = FALSE)

Arguments

scen

scenario

end_year

last full year for price simulation, set to 2045 by default

cru_cap

Boolean, defaults to TRUE

shock

if TRUE then a 4x price shock with time constant of one year is added on Jan 1 2030

Value

a eight-column dataframe of load_profiles,price, tariff_plan and HMM state

Details

Regulated network charges are ToU dependent and yeartime dependent. Charges at yeartime in a scenario are obtained from peak_network_charge_fun(scen,yeartime) etc. These are supposed to include supplier uplift, and other systems charges on top of the regulated network TUoS and DUos network charge. The model assumes that wholesale prices + VAT are passed through.

set_prices() is meant to to run at the beginning of each ABM run.

Optionally, the CRU wholesale price cap (currently 0.5euro/kWh) can be turned off.

Flat and ToU tariffs can be set retrospectively or based on forecast prices. Probably, retrospective price setting is best.

Examples

set_prices(sD)
#> # A tibble: 710,064 × 8
#>    datetime                  lp1       lp2       lp3       lp4 price tariff_plan
#>    <dttm>                  <dbl>     <dbl>     <dbl>     <dbl> <dbl> <chr>      
#>  1 2019-01-01 00:00:00 0.000100  0.000134  0.0000935 0.000124  0.229 flat       
#>  2 2019-01-01 01:00:00 0.0000788 0.000134  0.0000795 0.000112  0.229 flat       
#>  3 2019-01-01 02:00:00 0.0000733 0.000136  0.0000817 0.0000836 0.229 flat       
#>  4 2019-01-01 03:00:00 0.0000726 0.000139  0.0000656 0.0000784 0.229 flat       
#>  5 2019-01-01 04:00:00 0.0000741 0.000138  0.0000589 0.0000930 0.229 flat       
#>  6 2019-01-01 05:00:00 0.0000569 0.000131  0.0000541 0.0000872 0.229 flat       
#>  7 2019-01-01 06:00:00 0.0000628 0.000125  0.0000669 0.0000983 0.229 flat       
#>  8 2019-01-01 07:00:00 0.0000719 0.000112  0.0000705 0.0000952 0.229 flat       
#>  9 2019-01-01 08:00:00 0.0000815 0.0000872 0.0000873 0.0000952 0.229 flat       
#> 10 2019-01-01 09:00:00 0.0000943 0.000103  0.000120  0.000125  0.229 flat       
#> # ℹ 710,054 more rows
#> # ℹ 1 more variable: hmm_state <int>