survey_bills_to_kwh

a function that converts highest and lower bi-monthly bills from survey to daily D_max and D_min assuming a seasonal demand lag_D. D_max and D_min are used to estimate the financial return on pv bess investment. Missing data are imputed by default. An issue that needs to be addressed is "level pay". Stochastic.

survey_bills_to_kwh(data_in, lag_D = 30)

Arguments

data_in

survey data e.g. dep_survey

lag_D

seasonal lag in demand, default 30 days.

Examples

survey_bills_to_kwh(dep_survey,lag_D=30)
#> # A tibble: 1,225 × 11
#> # Rowwise: 
#>    serial   q14   q15    qi highest_bill lowest_bill lowest_kwh highest_kwh
#>     <dbl> <dbl> <dbl> <dbl>        <dbl>       <dbl>      <dbl>       <dbl>
#>  1     71     5     3     2         275         175        469.        738.
#>  2     72    13     2     5         278.        125        335.        747.
#>  3     74     4     2     5         225         125        335.        604.
#>  4     75    12    10     5         800         525       1408.       2146.
#>  5     76     6     1     3         325          50        134.        872.
#>  6     77     9     2     5         475         125        335.       1274.
#>  7     79     5    13     5         275         174.       466.        738.
#>  8     81     6     3     3         325         175        469.        872.
#>  9     82     6     5     5         325         275        738.        872.
#> 10     83    13    13     1         171.        110.       294.        459.
#> # ℹ 1,215 more rows
#> # ℹ 3 more variables: D_max <dbl>, D_min <dbl>, kWh <dbl>